created sub-components for each column
